throw error -> next(error)

+++ src/controllers/Category.controller.ts
@@ -1,4 +1,5 @@
-import { Request, Response } from 'express';
+import validator from "validator";
+import { NextFunction, Request, Response } from 'express';
 
 // User model
 // import Product from '../models/Product.model';
@@ -23,19 +24,18 @@ class CategoryController {
         this.CategoriesRepository = repository;
     };
 
-    public readonly getAllCategories = async (_req: Request, res: Response): Promise<Response> => {
+    public readonly getAllCategories = async (_req: Request, res: Response, next: NextFunction): Promise<Response | void> => {
         try {
             const categories = await this.CategoriesRepository.getAllCategories();
             return res.status(200).send({ categories });
         } catch (error: unknown) {
-            res.sendStatus(500);
-            throw error;
+            next(error);
         }
     };
 
-    public readonly getCategoryInfo = async (req: Request, res: Response): Promise<Response> => {
+    public readonly getCategoryInfo = async (req: Request, res: Response, next: NextFunction): Promise<Response | void> => {
         const id: string | undefined = req.params?.id;
-        if (!id) return res.sendStatus(400);
+        if (!id || !validator.isUUID(id)) return res.sendStatus(400);
 
         try {
             const category = await this.CategoriesRepository.getCategoryById(id);
@@ -43,12 +43,11 @@ class CategoryController {
 
             return res.status(200).send({ category });
         } catch (error: unknown) {
-            res.sendStatus(500);
-            throw error;
+            next(error);
         }
     };
     
-    public readonly addNewCategory = async (req: Request, res: Response): Promise<Response> => {
+    public readonly addNewCategory = async (req: Request, res: Response, next: NextFunction): Promise<Response | void> => {
         const newCategory: CategoryCreate | undefined = req.body;
         if (!newCategory) return res.sendStatus(400);
 
@@ -58,15 +57,14 @@ class CategoryController {
 
             return res.status(201).send({ ...category.get() });
         } catch (error: unknown) {
-            res.sendStatus(500);
-            throw error;
+            next(error);
         }
     };
 
-    public readonly editCategory = async (req: Request, res: Response): Promise<Response> => {
+    public readonly editCategory = async (req: Request, res: Response, next: NextFunction): Promise<Response | void> => {
         const id: string | undefined = req.params?.id;
         const newCategoryData: CategoryEdit | undefined = req.body;
-        if ( !id || !newCategoryData) return res.sendStatus(400);
+        if ( !id || !validator.isUUID(id) || !newCategoryData) return res.sendStatus(400);
 
         try {
             const category = await this.CategoriesRepository.updateCategory(id, newCategoryData);
@@ -74,23 +72,21 @@ class CategoryController {
 
             return res.status(200).send({ ...category.get() });
         } catch (error: unknown) {
-            res.sendStatus(500);
-            throw error;
+            next(error);
         }
     };
 
-    public readonly deletecategory = async (req: Request, res: Response): Promise<Response> => {
+    public readonly deletecategory = async (req: Request, res: Response, next: NextFunction): Promise<Response | void> => {
         try {
             const id: string | undefined = req.params?.id;
-            if(!id) return res.sendStatus(400);
+            if(!id || !validator.isUUID(id)) return res.sendStatus(400);
 
             const result = await this.CategoriesRepository.deleteCategory( id );
             if(!result) return res.sendStatus(404);
 
             return res.sendStatus(204);
         } catch (error: any) {
-            res.sendStatus(500);
-            throw new Error(error);
+            next(error);
         }  
     };
 };
