+++ src/controllers/SuperAdmin.controller.ts
@@ -1,5 +1,5 @@
-import { Request, Response } from 'express';
-import { ValidationError } from 'sequelize';
+import validator from "validator";
+import { NextFunction, Request, Response } from 'express';
 
 // class
 import UserController from './User.controller';
@@ -18,24 +18,22 @@ class SuperAdminController extends UserController {
     /**
      * GetAllUserInfo
      */
-    public readonly getAllUserInfo = async (_req: Request, res: Response): Promise<Response> => {
+    public readonly getAllUserInfo = async (_req: Request, res: Response, next: NextFunction): Promise<Response | void> => {
         try {
             const result = await this.UsersRepository.getAllUsers();
 
             return res.send({ users: result }).status(200);
         } catch (error: unknown) {  
-            if (error instanceof ValidationError) res.sendStatus(400);
-            else res.sendStatus(500);
-            throw error;
+            next(error);
         }
     };
 
     /**
      * GetUserInfo
      */
-     public readonly getUserInfo = async (req: Request, res: Response): Promise<Response> => {
+     public readonly getUserInfo = async (req: Request, res: Response, next: NextFunction): Promise<Response | void> => {
         const id: string | undefined = req.params.id;
-        if (!id) return res.sendStatus(404);
+        if (!id || !validator.isUUID(id)) return res.sendStatus(400);
 
         try {
             const user = await this.UsersRepository.getUserById( id );
@@ -43,8 +41,7 @@ class SuperAdminController extends UserController {
 
             return res.send( user ).status(200);
         } catch (error: unknown) {  
-            res.sendStatus(500);
-            throw error;
+            next(error);
         }
     };
 
@@ -52,7 +49,7 @@ class SuperAdminController extends UserController {
     /**
      * AddNewUser
      */
-    public readonly addNewUser = async (req: Request, res: Response): Promise<Response> => {
+    public readonly addNewUser = async (req: Request, res: Response, next: NextFunction): Promise<Response | void> => {
         try {
             const user: UserCreate | undefined = req.body;
             if(!user) return res.sendStatus(400);
@@ -62,18 +59,17 @@ class SuperAdminController extends UserController {
 
             return res.send({ user: result }).status(201);
         } catch (error: unknown) {
-            res.sendStatus(500);
-            throw error;
+            next(error);
         }  
     };
 
     /**
      * EditUser
      */
-     public readonly editUser = async (req: Request, res: Response): Promise<Response> => {
+     public readonly editUser = async (req: Request, res: Response, next: NextFunction): Promise<Response | void> => {
         try {
             const id: string | undefined = req.params?.id;
-            if(!id) return res.sendStatus(400);
+            if(!id || !validator.isUUID(id)) return res.sendStatus(400);
 
             const newUser: UserEdit = req.body;
             if(!newUser) return res.sendStatus(400);
@@ -83,26 +79,24 @@ class SuperAdminController extends UserController {
 
             return res.send({ ...user }).status(200);
         } catch (error: unknown) {
-            res.sendStatus(500);
-            throw error;
+            next(error);
         }  
     };
 
     /**
      * DeleteUser
      */
-     public readonly deleteUser = async (req: Request, res: Response): Promise<Response> => {
+     public readonly deleteUser = async (req: Request, res: Response, next: NextFunction): Promise<Response | void> => {
         try {
             const id: string | undefined = req.params?.id;
-            if(!id) return res.sendStatus(400);
+            if(!id || !validator.isUUID(id)) return res.sendStatus(400);
 
             const result = await this.UsersRepository.deleteUser( id );
             if(!result) return res.sendStatus(404);
 
             return res.sendStatus(204);
         } catch (error: unknown) {
-            res.sendStatus(500);
-            throw error;
+            next(error);
         }  
     };
 };
